Compatibility commit gcc version 12.2.0 Debian 12
* C++ 17 compliance and compatibility according to gcc 12.2
* matplotlib headers removed when they are not used
* in many tools there have been these
* CTIOL and CTMAL updated to gcc version as well
* minor code edits

* dentk-framecalc1
*    std::transform in many routines instead of for loops though frames
*    might increase efficacy

+++ src/dentk-svd.cpp
@@ -171,7 +171,7 @@ int main(int argc, char* argv[])
             lda = n;
             ldu = std::min(m, n);
             ldvt = n;
-            if(a.max_s_vals < ldu && a.max_s_vals > 0)
+            if((int)a.max_s_vals < ldu && a.max_s_vals > 0)
             {
                 range = 'I';
                 il = 1;
